+++ packages/packemon/src/index.ts
@@ -15,7 +15,6 @@ export * from './commands/Pack';
 export * from './commands/Scaffold';
 export * from './commands/Validate';
 export * from './commands/Watch';
-export * from './Config';
 export * from './constants';
 export * from './Package';
 export * from './PackageValidator';
